renamed deleteTestResultRecord function to deleteTestResult
