+++ serverless.yml
@@ -47,8 +47,8 @@ functions:
           path: test-result/retry
           method: get
 
-  deleteTestResultRecord:
-    handler: src/functions/deleteTestResultRecord/framework/handler.handler
+  deleteTestResult:
+    handler: src/functions/deleteTestResult/framework/handler.handler
     events:
       - http:
           path: test-result/delete
