+++ src/functions/deleteTestResult/framework/database/__tests__/query-builder.spec.ts
