+++ src/functions/deleteTestResult/framework/database/query-builder.ts
